Update MeshBasicMaterial type definition
A mesh can optionally have a `map` but right now this doesn't allow us to remove a material's map by setting it to `undefined`. As far as I understand, that's how you remove a map from an existing material. Unless there's another way to delete a map, I believe the typing info should be updated to allow this way.